src/app/google-maps/google-maps.component.ts
import { Component, OnInit } from '@angular/core';

@Component({
  selector: 'app-google-maps',
  templateUrl: './google-maps.component.html',
  styleUrls: ['./google-maps.component.css'],
})
export class GoogleMapsComponent implements OnInit {
  lat: number = 43.683334;
  lng: number = -79.76667;
  zoom: number = 15;
  locationChosen = false;

  constructor() {}

  ngOnInit() {}

  onChoseLocation(event) {
    console.log(event);
    this.lat = event.coords.lat;
    this.lng = event.coords.lng;
    this.locationChosen = true;
  }
}
